ltp: syscalls: fcntl: Clarify default pipe size for previleged user
The test doesn't set the pipe's size so we are actually checking the
default size. In the case of a privileged user this is determined by
PIPE_DEF_BUFFERS number of pages in the kernel.

User the kernel_page_size() instead of the userspace page size when
checking the pipe size on Android's x86_64 16KiB page size emulators.

This is effectively a no-op change on all other targets.

Bug: 366098040
Test: atest vts_ltp_test_x86_64:syscalls.fcntl05_64bit#syscalls.fcntl05_64bit
Change-Id: I7eb8839d595fd4ff96e5e19bca91e1228aba9759
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>