ltp: security: dirtyc0w_shmem: Fix backing file size
Android emulates a userspace 16KiB page size on x86_64 emulators.
The kernel however remains 4KiB and page faults are therefore handled
in 4KiB granulatiry.

A side effect of this is that there is a discrepancy in filemap faults
which usually allows faulting off the end of a file until the next page
boundary. In the userspace this page boundary is until the next 16KiB
aligned address, however for the kernel the next page boudnary is until
the next 4KiB aligned address. To solve this issue an anonymous mapping
is inserted by the emulator's kernel to cover the range between the
next 4KiB boundary after the file end and the next 16KiB page boundary.
This causes the uffd register operation to fail for minor faults since
the backing VMA is no longer shmem but anonymous.

Work around this by making the backing shmem file just large enough to
partially cover the last 4KiB page before the next 16KiB boundary after
the file's end.

Bug: 366098040
Test: atest vts_ltp_test_x86_64:cve.cve-2022-2590_64bit#cve.cve-2022-2590_64bit
Change-Id: Ie6e9c77870366a34b8304d87b71af71abb7f4688
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>