ltp: syscalls: epoll_wait: Work around edge triggered semantics
The documentation says that it "might" and "probably" won't recieve
a new event for exisitng data. So we can't be guaranteed that it
won't.

       If the rfd file descriptor has been added to the epoll interface
       using the EPOLLET (edge-triggered) flag, the call to
       epoll_wait(2) done in step 5 will **probably** hang despite the
       available data still present in the file input buffer; meanwhile
       the remote peer **might** be expecting a response based on the data
       it already sent.

[1] https://man7.org/linux/man-pages/man7/epoll.7.html

It does return a new event in the case the pipe size is larger than
a single page.

Update the test to use a single kernel page for the pipe and write
sizes.

Bug: 366098040
Test: atest vts_ltp_test_x86_64:syscalls.epoll_wait06_64bit#syscalls.epoll_wait06_64bit
Change-Id: Ie33174adaa2473ce0db4617ff3e99c06d32c8648
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>