ANDROID: Merge LTP from upstream tag '20250530'.
This commit updates the Linux Test Project (LTP) to the latest upstream release, version 20250530.
The commits from the upstream repository have been squashed and merged as a single commit into this code base.
Removed the `tools/kirk` directory for resolving the conflict, and introduce no other changes. This directory exists upstream but not already exist in the Android repository.

Upstream Source:
* Repo: Linux Test Project (LTP)
* Release Tag: 20250530
* URL: https://github.com/linux-test-project/ltp/releases/tag/20250530


Test: Tested on another change in the submitted changes
Bug: 422759691
Signed-off-by: Darren Chang <chihsheng@google.com>
Change-Id: Ifad04ade0e737f021956c07680fea52216a463c8